Backend/requirements.txt
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
tensorflow>=2.13.0
Pillow>=10.0.0
numpy>=1.24.0
